GH-38 Serve Service Worker from the correct location
Previously, the service worker was symlinked to the root of the `public`
directory to grant it the desired scope.
This was a misguided decision. The correct solution is to use the
[`Service-Worker-Allowed`](https://developer.mozilla.org/en-US/docs/Web/API/ServiceWorkerContainer/register#scope)
header, as described in the MDN documentation.

This commit introduces that change. It also improves the developer experience by
using:

- Reitit router
- LambdaIsland Hiccup for response generation